Split pfPasswordStore into platform-specific files.
