+++ Sources/Plasma/FeatureLib/pfPasswordStore/CMakeLists.txt
@@ -12,6 +12,24 @@ set(pfPasswordStore_SOURCES
     pfPasswordStore.cpp
 )
 
+if(WIN32 AND NOT CYGWIN)
+    set(pfPasswordStore_SOURCES ${pfPasswordStore_SOURCES}
+        pfPasswordStore_Win.cpp
+    )
+endif(WIN32 AND NOT CYGWIN)
+
+if(UNIX)
+    set(pfPasswordStore_SOURCES ${pfPasswordStore_SOURCES}
+        pfPasswordStore_Unix.cpp
+    )
+endif(UNIX)
+
+if(APPLE)
+    set(pfPasswordStore_SOURCES ${pfPasswordStore_SOURCES}
+        pfPasswordStore_Mac.cpp
+    )
+endif(APPLE)
+
 add_library(pfPasswordStore STATIC ${pfPasswordStore_HEADERS} ${pfPasswordStore_SOURCES})
 target_link_libraries(pfPasswordStore CoreLib plFile)
