+++ Sources/Plasma/FeatureLib/pfPasswordStore/pfPasswordStore.cpp
@@ -142,112 +142,3 @@ bool pfFilePasswordStore::SetPassword(const plString& username, const plString& 
 
     return false;
 }
-
-
-
-#ifdef HS_BUILD_FOR_WIN32
-#include "hsWindows.h"
-#include <wincred.h>
-
-/*****************************************************************************
- ** pfWin32PasswordStore                                                    **
- *****************************************************************************/
-const plString pfWin32PasswordStore::GetPassword(const plString& username)
-{
-    PCREDENTIALW credential;
-    plString target = plString::Format("%s__%s", plProduct::UUID(), username.c_str());
-    plString password = plString::Null;
-
-    if (!CredReadW(target.ToWchar().GetData(), CRED_TYPE_GENERIC, 0, &credential)) {
-        return password;
-    }
-
-    password = plString::FromUtf8(reinterpret_cast<const char *>(credential->CredentialBlob), credential->CredentialBlobSize);
-
-    memset(credential->CredentialBlob, 0, credential->CredentialBlobSize);
-    CredFree(credential);
-
-    return password;
-}
-
-
-bool pfWin32PasswordStore::SetPassword(const plString& username, const plString& password)
-{
-    CREDENTIALW credential;
-    plString target = plString::Format("%s__%s", plProduct::UUID(), username.c_str());
-
-    if (password.IsNull()) {
-        if (CredDeleteW(target.ToWchar().GetData(), CRED_TYPE_GENERIC, 0)) {
-            return true;
-        }
-        return false;
-    }
-
-    plStringBuffer<wchar_t> tbuff = target.ToWchar();
-    plStringBuffer<char> pbuff = password.ToUtf8();
-    plStringBuffer<wchar_t> ubuff = username.ToWchar();
-
-    memset(&credential, 0, sizeof(CREDENTIALW));
-    credential.Type = CRED_TYPE_GENERIC;
-    credential.TargetName = (LPWSTR)tbuff.GetData();
-    credential.CredentialBlobSize = pbuff.GetSize();
-    credential.CredentialBlob = (LPBYTE)pbuff.GetData();
-    credential.Persist = CRED_PERSIST_LOCAL_MACHINE;
-    credential.UserName = (LPWSTR)ubuff.GetData();
-
-    if (!CredWriteW(&credential, 0)) {
-        return false;
-    }
-
-    return true;
-}
-#endif
-
-
-
-#ifdef HS_BUILD_FOR_OSX
-#include <Security/Security.h>
-
-/*****************************************************************************
- ** pfMacPasswordStore                                                      **
- *****************************************************************************/
-const plString pfMacPasswordStore::GetPassword(const plString& username)
-{
-    plString service = plProduct::UUID();
-
-    void* passwd = nullptr;
-    uint32_t passwd_len = 0;
-
-    if (SecKeychainFindGenericPassword(nullptr,
-                                       service.GetSize(),
-                                       service.c_str(),
-                                       username.GetSize(),
-                                       username.c_str(),
-                                       &passwd_len,
-                                       &passwd,
-                                       nullptr) != errSecSuccess)
-    {
-        return plString::Null;
-    }
-
-    plString ret(reinterpret_cast<const char*>(passwd), size_t(passwd_len));
-
-    SecKeychainItemFreeContent(nullptr, passwd);
-
-    return ret;
-}
-
-bool pfMacPasswordStore::SetPassword(const plString& username, const plString& password)
-{
-    plString service = plProduct::UUID();
-
-    return SecKeychainAddGenericPassword(nullptr,
-                                         service.GetSize(),
-                                         service.c_str(),
-                                         username.GetSize(),
-                                         username.c_str(),
-                                         password.GetSize(),
-                                         password.c_str(),
-                                         nullptr) == errSecSuccess;
-}
-#endif
