+++ Sources/Plasma/FeatureLib/pfPasswordStore/pfPasswordStore_Mac.cpp
@@ -1,0 +1,91 @@
+/*==LICENSE==*
+
+CyanWorlds.com Engine - MMOG client, server and tools
+Copyright (C) 2011  Cyan Worlds, Inc.
+
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+Additional permissions under GNU GPL version 3 section 7
+
+If you modify this Program, or any covered work, by linking or
+combining it with any of RAD Game Tools Bink SDK, Autodesk 3ds Max SDK,
+NVIDIA PhysX SDK, Microsoft DirectX SDK, OpenSSL library, Independent
+JPEG Group JPEG library, Microsoft Windows Media SDK, or Apple QuickTime SDK
+(or a modified version of those libraries),
+containing parts covered by the terms of the Bink SDK EULA, 3ds Max EULA,
+PhysX SDK EULA, DirectX SDK EULA, OpenSSL and SSLeay licenses, IJG
+JPEG Library README, Windows Media SDK EULA, or QuickTime SDK EULA, the
+licensors of this Program grant you additional
+permission to convey the resulting work. Corresponding Source for a
+non-source form of such a combination shall include the source code for
+the parts of OpenSSL and IJG JPEG Library used as well as that of the covered
+work.
+
+You can contact Cyan Worlds, Inc. by email legal@cyan.com
+ or by snail mail at:
+      Cyan Worlds, Inc.
+      14617 N Newport Hwy
+      Mead, WA   99021
+
+*==LICENSE==*/
+
+#include "pfPasswordStore.h"
+#include "pfPasswordStore_impl.h"
+
+#include "plProduct.h"
+
+#include <Security/Security.h>
+
+/*****************************************************************************
+ ** pfMacPasswordStore                                                      **
+ *****************************************************************************/
+const plString pfMacPasswordStore::GetPassword(const plString& username)
+{
+    plString service = plProduct::UUID();
+
+    void* passwd = nullptr;
+    uint32_t passwd_len = 0;
+
+    if (SecKeychainFindGenericPassword(nullptr,
+                                       service.GetSize(),
+                                       service.c_str(),
+                                       username.GetSize(),
+                                       username.c_str(),
+                                       &passwd_len,
+                                       &passwd,
+                                       nullptr) != errSecSuccess)
+    {
+        return plString::Null;
+    }
+
+    plString ret(reinterpret_cast<const char*>(passwd), size_t(passwd_len));
+
+    SecKeychainItemFreeContent(nullptr, passwd);
+
+    return ret;
+}
+
+bool pfMacPasswordStore::SetPassword(const plString& username, const plString& password)
+{
+    plString service = plProduct::UUID();
+
+    return SecKeychainAddGenericPassword(nullptr,
+                                         service.GetSize(),
+                                         service.c_str(),
+                                         username.GetSize(),
+                                         username.c_str(),
+                                         password.GetSize(),
+                                         password.c_str(),
+                                         nullptr) == errSecSuccess;
+}
